pyproject.toml
[project]
name = "ai-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "fastapi>=0.110.0",
    "groq>=0.9.0",
    "gtts>=2.5.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.50",
    "numpy>=1.26.0",
    "openai>=1.0.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0.0",
    "scipy>=1.11.0",
    "uvicorn>=0.27.0",
]
